Trabalho 6/sistema-carga-ia/consumer_face/requirements.txt
pika==1.2.0
Pillow==9.4.0
python-dotenv==0.21.0
opencv-python-headless==4.7.0.72
deepface==0.0.79
numpy==1.24.2
